requirements.txt
streamlit==1.38.0
pandas==2.2.2
plotly==5.23.0
requests==2.32.3
beautifulsoup4==4.12.3
python-dateutil==2.9.0.post0
